backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
gunicorn==20.1.0
python-dotenv==0.19.0
requests==2.26.0
pymongo==4.0.1
pandas==1.3.3
numpy==1.21.2
scikit-learn==1.0 